update ceil and add div && mod

+++ p1/mc1/avu2/ceil.cpp
@@ -1,11 +1,10 @@
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 #include "doctest.h"
-#include <iostream>
 
 // n = x + f
 // x >= 0 (dado 0 < f < 1 logo n < x < n+1)
 // ceil -> n < x < n+1
-int ceil(float x) {
+int ceil(int x) {
   int n = 0;
   while (true) {
     if (x < 0) {
